@@ -1,12 +1,13 @@
 module github.com/AzureAD/microsoft-authentication-library-for-go
 
-go 1.14
+go 1.17
 
 require (
-	github.com/golang-jwt/jwt v3.2.1+incompatible
+	github.com/golang-jwt/jwt/v4 v4.4.2
 	github.com/google/uuid v1.1.1
 	github.com/kylelemons/godebug v1.1.0
-	github.com/montanaflynn/stats v0.6.6 // indirect
+	github.com/montanaflynn/stats v0.6.6
 	github.com/pkg/browser v0.0.0-20210115035449-ce105d075bb4
-	golang.org/x/sys v0.0.0-20210124154548-22da62e12c0c // indirect
 )
+
+require golang.org/x/sys v0.0.0-20210124154548-22da62e12c0c // indirect
