@@ -19,6 +19,7 @@ import (
 	"time"
 
 	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/errors"
+	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/internal/exported"
 	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/internal/oauth/fake"
 	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/internal/oauth/ops/accesstokens"
 	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/internal/oauth/ops/authority"
@@ -66,6 +67,9 @@ func TestAuthCode(t *testing.T) {
 }
 
 func TestCredential(t *testing.T) {
+	callback := func(context.Context, exported.AssertionRequestOptions) (string, error) {
+		return "assertion", nil
+	}
 	tests := []struct {
 		desc       string
 		re         fake.ResolveEndpoints
@@ -79,8 +83,7 @@ func TestCredential(t *testing.T) {
 			re:   fake.ResolveEndpoints{Err: true},
 			at:   &fake.AccessTokens{},
 			cred: &accesstokens.Credential{
-				Assertion: "assertion",
-				Expires:   time.Now().Add(-5 * time.Minute),
+				AssertionCallback: callback,
 			},
 			err: true,
 		},
@@ -89,8 +92,7 @@ func TestCredential(t *testing.T) {
 			re:   fake.ResolveEndpoints{},
 			at:   &fake.AccessTokens{Err: true},
 			cred: &accesstokens.Credential{
-				Assertion: "assertion",
-				Expires:   time.Now().Add(-5 * time.Minute),
+				AssertionCallback: callback,
 			},
 			err: true,
 		},
@@ -99,9 +101,8 @@ func TestCredential(t *testing.T) {
 			re:   fake.ResolveEndpoints{},
 			at:   &fake.AccessTokens{Err: true},
 			cred: &accesstokens.Credential{
-				Assertion: "assertion",
-				Expires:   time.Now().Add(5 * time.Minute),
-				Cert:      &x509.Certificate{},
+				AssertionCallback: callback,
+				Cert:              &x509.Certificate{},
 				// Key is nil and causes token.SignedString(c.Key) to fail in Credential.JWT()
 			},
 			err: true,
@@ -111,8 +112,7 @@ func TestCredential(t *testing.T) {
 			re:   fake.ResolveEndpoints{},
 			at:   &fake.AccessTokens{Err: true},
 			cred: &accesstokens.Credential{
-				Assertion: "assertion",
-				Expires:   time.Now().Add(-5 * time.Minute),
+				AssertionCallback: callback,
 			},
 			err: true,
 		},
@@ -129,8 +129,7 @@ func TestCredential(t *testing.T) {
 			re:   fake.ResolveEndpoints{},
 			at:   &fake.AccessTokens{},
 			cred: &accesstokens.Credential{
-				Assertion: "assertion",
-				Expires:   time.Now().Add(-5 * time.Minute),
+				AssertionCallback: callback,
 			},
 		},
 	}
